move package-specific types out of utils
